+++ cmd/bricksllm/main.go
@@ -12,6 +12,7 @@ import (
 	auth "github.com/bricks-cloud/bricksllm/internal/authenticator"
 	"github.com/bricks-cloud/bricksllm/internal/cache"
 	"github.com/bricks-cloud/bricksllm/internal/config"
+	"github.com/bricks-cloud/bricksllm/internal/encryptor"
 	"github.com/bricks-cloud/bricksllm/internal/logger/zap"
 	"github.com/bricks-cloud/bricksllm/internal/manager"
 	"github.com/bricks-cloud/bricksllm/internal/message"
@@ -173,22 +174,25 @@ func main() {
 	}
 	rMemStore.Listen()
 
-	rateLimitRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       0,
-	})
+	defaultRedisOption := func(cfg *config.Config, dbIndex int) *redis.Options {
+
+		options := &redis.Options{
+			Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
+			Password: cfg.RedisPassword,
+			DB:       cfg.RedisDBStartIndex + dbIndex,
+		}
+
+		return options
+	}
+
+	rateLimitRedisCache := redis.NewClient(defaultRedisOption(cfg, 0))
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
 	if err := rateLimitRedisCache.Ping(ctx).Err(); err != nil {
 		log.Sugar().Fatalf("error connecting to rate limit redis cache: %v", err)
 	}
 
-	costLimitRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       1,
-	})
+	costLimitRedisCache := redis.NewClient(defaultRedisOption(cfg, 1))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -196,11 +200,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to cost limit redis cache: %v", err)
 	}
 
-	costRedisStorage := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       2,
-	})
+	costRedisStorage := redis.NewClient(defaultRedisOption(cfg, 2))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -208,11 +208,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to cost limit redis storage: %v", err)
 	}
 
-	apiRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       3,
-	})
+	apiRedisCache := redis.NewClient(defaultRedisOption(cfg, 3))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -220,11 +216,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to api redis cache: %v", err)
 	}
 
-	accessRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       4,
-	})
+	accessRedisCache := redis.NewClient(defaultRedisOption(cfg, 4))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -232,11 +224,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to api redis cache: %v", err)
 	}
 
-	userRateLimitRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       5,
-	})
+	userRateLimitRedisCache := redis.NewClient(defaultRedisOption(cfg, 5))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -244,11 +232,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to user rate limit redis cache: %v", err)
 	}
 
-	userCostLimitRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       6,
-	})
+	userCostLimitRedisCache := redis.NewClient(defaultRedisOption(cfg, 6))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -256,11 +240,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to user cost limit redis cache: %v", err)
 	}
 
-	userCostRedisStorage := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       7,
-	})
+	userCostRedisStorage := redis.NewClient(defaultRedisOption(cfg, 7))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -268,11 +248,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to user cost redis cache: %v", err)
 	}
 
-	userAccessRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       8,
-	})
+	userAccessRedisCache := redis.NewClient(defaultRedisOption(cfg, 8))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -280,11 +256,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to user access redis storage: %v", err)
 	}
 
-	providerSettingsRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       9,
-	})
+	providerSettingsRedisCache := redis.NewClient(defaultRedisOption(cfg, 9))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -292,11 +264,7 @@ func main() {
 		log.Sugar().Fatalf("error connecting to provider settings redis storage: %v", err)
 	}
 
-	keysRedisCache := redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
-		Password: cfg.RedisPassword,
-		DB:       10,
-	})
+	keysRedisCache := redis.NewClient(defaultRedisOption(cfg, 10))
 
 	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -318,11 +286,16 @@ func main() {
 	psCache := redisStorage.NewProviderSettingsCache(providerSettingsRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
 	keysCache := redisStorage.NewKeysCache(keysRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
 
+	encryptor, err := encryptor.NewEncryptor(cfg.DecryptionEndpoint, cfg.EncryptionEndpoint, cfg.EnableEncrytion, cfg.EncryptionTimeout, cfg.Audience)
+	if cfg.EnableEncrytion && err != nil {
+		log.Sugar().Fatalf("error creating encryption client: %v", err)
+	}
 	v := validator.NewValidator(costLimitCache, rateLimitCache, costStorage)
+
 
 	m := manager.NewManager(store, costLimitCache, rateLimitCache, accessCache, keysCache)
 	krm := manager.NewReportingManager(costStorage, store, store, v)
-	psm := manager.NewProviderSettingsManager(store, psCache)
+	psm := manager.NewProviderSettingsManager(store, psCache, encryptor)
 	cpm := manager.NewCustomProvidersManager(store, cpMemStore)
 	rm := manager.NewRouteManager(store, store, rMemStore, psm)
 	pm := manager.NewPolicyManager(store, rMemStore)
@@ -359,7 +332,7 @@ func main() {
 
 	rec := recorder.NewRecorder(costStorage, userCostStorage, costLimitCache, userCostLimitCache, ce, store)
 	rlm := manager.NewRateLimitManager(rateLimitCache, userRateLimitCache)
-	a := auth.NewAuthenticator(psm, m, rm, store)
+	a := auth.NewAuthenticator(psm, m, rm, store, encryptor)
 
 	c := cache.NewCache(apiCache)