+++ internal/manager/route.go
@@ -101,8 +101,10 @@ func checkModelValidity(provider, model string) bool {
 
 var (
 	azureSupportedModels = []string{
+		"gpt-4o-2024-08-26",
 		"gpt-4o-2024-05-13",
 		"gpt-4o",
+		"gpt-4o-mini",
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-0125-preview",
 		"gpt-4-vision-preview",
@@ -115,8 +117,6 @@ var (
 		"gpt-4-32k-0613",
 		"gpt-4-32k-0314",
 		"gpt-4-turbo",
-		"gpt-4o-2024-05-13",
-		"gpt-4o",
 		"gpt-35-turbo",
 		"gpt-35-turbo-0125",
 		"gpt-35-turbo-1106",
@@ -129,8 +129,10 @@ var (
 	}
 
 	openaiSupportedModels = []string{
+		"gpt-4o-2024-08-06",
 		"gpt-4o-2024-05-13",
 		"gpt-4o",
+		"gpt-4o-mini",
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-0125-preview",
 		"gpt-4-vision-preview",
@@ -146,8 +148,6 @@ var (
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-1106-preview",
 		"gpt-4-turbo-preview",
-		"gpt-4o-2024-05-13",
-		"gpt-4o",
 		"gpt-3.5-turbo",
 		"gpt-3.5-turbo-0125",
 		"gpt-3.5-turbo-1106",
@@ -160,8 +160,10 @@ var (
 	}
 
 	supportedModels = []string{
+		"gpt-4o-2024-08-06",
 		"gpt-4o-2024-05-13",
 		"gpt-4o",
+		"gpt-4o-mini",
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-0125-preview",
 		"gpt-4-vision-preview",
@@ -177,8 +179,6 @@ var (
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-1106-preview",
 		"gpt-4-turbo-preview",
-		"gpt-4o-2024-05-13",
-		"gpt-4o",
 		"gpt-35-turbo",
 		"gpt-35-turbo-0125",
 		"gpt-35-turbo-1106",
@@ -215,8 +215,10 @@ var (
 		"gpt-35-turbo-0613",
 		"gpt-35-turbo-16k",
 		"gpt-35-turbo-16k-0613",
+		"gpt-4o-2024-08-06",
 		"gpt-4o-2024-05-13",
 		"gpt-4o",
+		"gpt-4o-mini",
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-0125-preview",
 		"gpt-4-vision-preview",
@@ -232,8 +234,6 @@ var (
 		"gpt-4-turbo-2024-04-09",
 		"gpt-4-1106-preview",
 		"gpt-4-turbo-preview",
-		"gpt-4o-2024-05-13",
-		"gpt-4o",
 		"gpt-3.5-turbo",
 		"gpt-3.5-turbo-0125",
 		"gpt-3.5-turbo-1106",